feat: README.md Task 1 and 2 are complete
The menu is fully functional and pops into existence when the hamburger is clicked. It may need to be jazzed up with some fun animation later.

The articles now have an 'expand' text that expands the article to 400px when clicked. It doesn't show the full article and should expand to a more appropriate height with some fun animation at a later date.