feat: Adds animation to sidebar navigation
Add a sidebar navigation that slides in from the side with a slight bounce at the end and reverses back off the edge. This function has a side effect that hasn't been taken care of yet where the time it takes to back off the screen is lengthened by the amount of time it spends on screen. I haven't yet been able to figure out how to control the time parameters such that that doesn't happen.