completed greensock menu sliding functionality
Menu now slides back in whenever the content area is clicked only if the
menu is already out.   Accomplished using an if statement in the event
listener on the click event of the area container.  The if statement checked
if the style.transform property had a matrix value indicating that the menu
was already out and only then slid the menu back if clicked.